refactor: isolate DIESEL transaction integration
