+++ src/core/counterparty/pendingChange.ts
@@ -19,13 +19,13 @@
  */
 
 import { decodeDieselMintScript } from '@/core/alkanes/diesel';
-import { normalizeAddressForComparison } from '@/core/bitcoin/address';
-import { parseRawTransactionLocally } from '@/core/bitcoin/localTransactionParse';
 import {
   type PendingDieselUtxo,
-  recordPendingChange,
   recordPendingDieselUtxo,
-} from '@/core/bitcoin/spentUtxoCache';
+} from '@/core/alkanes/pendingDieselUtxos';
+import { normalizeAddressForComparison } from '@/core/bitcoin/address';
+import { parseRawTransactionLocally } from '@/core/bitcoin/localTransactionParse';
+import { recordPendingChange } from '@/core/bitcoin/spentUtxoCache';
 import { unpackCounterpartyMessage } from '@/core/counterparty/unpack';
 import { extractPayloadFromOutputs } from '@/core/counterparty/unpack/opReturn';
